Prevented unautarized access to admin panel.
Prevented unauthorized access to the admin panel. Installed
express-session to create a session and sesssion id which is stored in
cookies. The session expires when browser closes.

Also added a logout functionality, however it was commented out for now because it logs out
on page refresh also. It is only supposed to logout when the tab closes.
(Should be fixed in the future)

The sessions are stored in server memory (can also be stored in a DB),
hence they get deleted when server restarts.