Intalled EJS and make the NFC links work
Installed EJS to render the files from server side using variables. Also
set the ENV (enviroment variable) to production / development as needed
(change from .env)

Using routing parameters and EJS made the links work. The review link in
the page is rendered differntly based on wether or not a placeId is
given. E.G If placeId is given then it works normally, if placeId is not
given but the url for that page exists in the database then the button
is just disabled. If there is not url for the page in database then it
returns CANNOT GET error.

In DBService.js the link creation was changed. The first part of the URL
is not hardcoded anymore. It takes whatever value is given in the .env
folder. So that if testing in localhost it retruns new links with the
localhost:5000 beginning.

Also the confirm edit was edited, in the last commit the changes weren't
shown in the UI when the confirm edit button was pressed. Now it works
:)